Update Guava to 11.0.2.
Whilst it is possible for a change such as this to be breaking, the version bump in itself is very small and Guava only removes APIs after 18 months deprecation. Given the proliferation of servers such as MCPC bundling Guava versions of 14.0 upwards as well as the fact that developers should steer clear away from using deprecated APIs, it is very unlikely this change will cause any issues. Furthermore this version brings multiple bug fixes and represents the minimum version required by Minecraft itself.