Revert Changes to Player Movement
Once and for all.

Thanks to Thinkofdeath & the users who put up with these changes, it was for a good cause, however at this point in time we do not feel it is possible to correct the large number of exploits present in player movement without the cooperation of Mojang. We will be seeking their cooperation, however in the mean time it is recommended that users install an anticheat plugin such as NoCheat+ to ensure some level of protection.

This will fix some strange behaviour users have been seeing in relation to player movements.