Add support for Minecraft 1.8.x
This commit allows Spigot to support Minecraft clients both of versions 1.7.x and of 1.8.x.

*** It is imperative that backups be taken before installing this build on your server! ***

Please note that this commit is NOT a fully functioning 1.8 server and does not contain all aspects of 1.8 support.

Many months of time and effort have gone into ensuring that this patch is as complete, bug free and reliable as possible, however it is highly likely there will be as of yet undiscovered bugs and glitches.

Please report any such issues to the bug report forum, stating your client, server, and if applicable BungeeCord versions.

*** Any bug reports not running the very latest versions of these softwares will be ignored ***