Ensure an entity that needs to be added to chunk ticks
It's possible an entity could be marked for deactivation while its in a
"needs to change chunks" state. This can cause it to then save to the world in the wrong chunk,
causing the "saved in wrong location" warnings.

This checks the boolean for isAddedToChunk and always ticks if its false.