Fix Fireworks on edge of unloaded chunks issue
This was something I fixed in the past but it ended up getting back in it seems.

Fireworks absolutely can not be skipped due to being on the edge of unloaded chunks,
or else they will stick around forever and never despawn until players move near.

On some servers, fireworks are infinite and redstone machines will build up tons of fireworks,
causing crushing client lag and server lag.

This resolves that by letting fireworks continue to move and detonate.