Rewrite of the player chunk management Loading of all chunks within a player region is not done within a single tick anymore, but instead they are queued and processed over a period of several ticks, depending on how much can be done without exceeding the 50ms limit of one tick and without exceeding the average limit of chunks that can be generated in a single tick.
On the settings maxChunkGenerationRate:
A value of 1.0 means that on average one chunk is allowed to be generated per tick

The process of creating MapChunk and MapChunkBulk packets is offloaded now to a pool of threads. These already do the chunk orebfuscation for the single chunks packet as well, but not yet for bulk ones.

There's also a new chunk coordinate comparator, which favors chunks in the players field of view (based on the yaw value) over those outside of it